+++ Tests/TaskRunnerTest.php
@@ -7,10 +7,17 @@ use Inspectioneering\TaskRunner\TaskRunner;
 
 class TaskRunnerTest extends \PHPUnit_Framework_TestCase
 {
-    public function testMissingConfigFileTaskThrowsException()
+    /**
+     * If the config contains an invalid/missing bootstrap file, it should throw an error.
+     */
+    public function testMissingBootstrapFileTaskThrowsException()
     {
+        $config = array(
+            'bootstrap' => 'path/to/some/fake/bootstrap/file.php',
+        );
+
         $this->setExpectedException(TaskException::class);
 
-        $runner = new TaskRunner('/path/to/some/fake/config/dir');
+        $runner = new TaskRunner($config);
     }
 }